+++ packages/plugin/convert/examples/simple/recourses/file.js
@@ -1,5 +1,3 @@
-// utils.mjs
-
 /**
  * Adds two numbers together.
  *