+++ packages/utils-media/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dovenv/utils-media
 
+## 1.5.16
+
+### Patch Changes
+
+- Update deps
+
+- Updated dependencies []:
+  - @dovenv/utils@1.5.16
+
 ## 1.5.15
 
 ### Patch Changes